refactor: domain 패키지의 내부 패키지를 변경
- game: 사다리 게임의 로직에 포함되는 주요 클래스가 위치함
- resource : 사다리 게임에 사용되는 데이터가 위치함
