+++ src/main/java/ladder/domain/resource/direction/DirectionGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain.direction;
+package ladder.domain.resource.direction;
 
 public interface DirectionGenerator {
