+++ src/main/java/ladder/view/OutputView.java
@@ -1,11 +1,11 @@
 package ladder.view;
 
 import java.util.List;
-import ladder.domain.direction.Direction;
-import ladder.domain.ladder.Ladder;
-import ladder.domain.line.Line;
-import ladder.domain.user.User;
-import ladder.domain.user.Users;
+import ladder.domain.resource.direction.Direction;
+import ladder.domain.resource.ladder.Ladder;
+import ladder.domain.resource.line.Line;
+import ladder.domain.resource.user.User;
+import ladder.domain.resource.user.Users;
 
 public class OutputView {
