+++ src/test/java/ladder/domain/resource/direction/DirectionRightGeneratorImpl.java
@@ -1,10 +1,10 @@
-package ladder.domain.direction;
+package ladder.domain.resource.direction;
 
-public class DirectionGeneratorImpl implements DirectionGenerator {
+public class DirectionRightGeneratorImpl implements DirectionGenerator {
 
     @Override
     public Direction generateInitialValue() {
-        return generateRightOrNeutral();
+        return Direction.getDirection(0);
     }
 
     @Override
