+++ src/test/java/ladder/domain/resource/line/LineTest.java
@@ -1,10 +1,10 @@
-package ladder.domain.line;
-
-import ladder.domain.direction.Direction;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+package ladder.domain.resource.line;
 
 import static org.assertj.core.api.Assertions.assertThat;
+
+import ladder.domain.resource.direction.Direction;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 class LineTest {
